Add the default theme for Ink.
The default theme is built to be suitable for all Beanbag products. It
provides:

* A color palette based on Review Board's standard palette
* Numerous palette design tokens that components can be built upon
* Standard units, fonts, icons, z-indexes, borders, and more.

This isn't complete, and many of exact colors are going to be altered as
time goes on, but this sets up the foundation for future development.

We're using Tabler.io for the icons (all MIT-licensed), including their
SVGs as data-URIs in the stylesheet. This gives us a standard set of
scalable icons we can use without any additional downloads.

Note that the theme does not set any styles on the actual page. As in,
it doesn't set the page background, font, headers, etc. It merely provides
styling for Ink's components. It's up to the consuming application to make
use of these. An "auto" style will be provided for setting some defaults.

Testing Done:
Made use of many of the variables within this in the development of
the new components.

Reviewed at https://reviews.reviewboard.org/r/13584/