Fix a bad variable name when injecting custom rules to schemas.
The old name, `@_override-rules`, was used in place of the new name,
`@_specialize-rules`. This is a trivial fix to use the correct name,
fixing the ability to apply specialized rules at the theme layer.

Testing Done:
Made use of this in in-progress components.

Reviewed at https://reviews.reviewboard.org/r/13624/