Produce cleaner readable and minified CSS files.
We now depend on `cleancss`, a tool that helps consolidate, clean, and
minify or beautify CSS files. It will take care of consolidating
redundant rules and media selectors, removing comments, using more terse
forms of CSS rules, and more.

This is used as part of our build pipeline. We produce an initial
`_build/ink.css` file, which then gets turned into a `lib/ink.css` that
is readable but optimized, and a `lib/ink.min.css` that is compact and
suited for browsers.

Testing Done:
Built the CSS. Compared the resulting versions to the source files,
verifying that nothing important was lost but that a lot of redundancy
was cleaned up.

Also verified that the `ink.css` is readable while `ink.min.css` is
minified.

Reviewed at https://reviews.reviewboard.org/r/13618/