Add a button group component.
Button groups are a container for multiple buttons, grouped together as
a single unit. They're useful when presenting a set of related actions,
and can be shown horizontally or stacked vertically.

Note that these should not be confused with toolbars, which are similar
in concept but would have a different presentation and capabilities.

Testing Done:
Unit tests pass.

Played with the different layouts and button type combinations in
Storybook.

Reviewed at https://reviews.reviewboard.org/r/13619/