Add a spinner component.
This is a very simple spinner, showing an operation is in progress (such
as content loading from a server). It uses a typical semi-circle image,
the one we have in Djblets, masked to the current color (which allows
some degree of customization).

By itself, this component does not provide any accessibility
information. Due to the various ways this can be used, that's an
exercise left to the consumer. Guidance is provided in the
documentation.

Testing Done:
Unit tests pass.

Played with the spinners in Storybook. These are basic so there isn't
much to test.

Reviewed at https://reviews.reviewboard.org/r/13620/