Add a Menu Button component.
The Menu Button is a button that displays a drop-down menu on hover,
click, or keyboard interaction. This can be used to select an action or
change settings for a piece of UI.

Menu Buttons can have a distinct Action Button, which will invoke an
action but will not show the drop-down menu. In this situation, the menu
is generally used for configuration or alternate (but related) actions.

Testing Done:
Tested the component thoroughly in Storybook.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/13644/