Add autofocus and arbitrary attribute support for buttons.
This introduces a new `autofocus` option/property for Ink.Button, and an
`attrs` for setting arbitrary attributes. This is useful for forms and
dialogs that need to go beyond the standard button options.

Testing Done:
Unit tests pass.

Made use of these in in-progress code.

Reviewed at https://reviews.reviewboard.org/r/13716/