Rename data-theme to data-ink-color-scheme.
The name `data-theme` was a bit too generic, and implied that color
schemes and themes were the same thing. That boxes us in in the future,
and could clash with other uses.

This change renames it to `data-ink-color-scheme`, which is far more
clear on purpose, and can be reused amongst other themes in the future,
if ones were introduced.

It also adds "system" and "high-contrast" as options in Storybook,

Testing Done:
Tested in Storybook and in Review Board with accompanying changes.

Reviewed at https://reviews.reviewboard.org/r/13726/