Flesh out accent styling, and use it for alerts.
This updates our accent styling colors for info, success, error,
warning, and highlight based on the alert colors, and adds outline
colors for each.

Alerts are updated to use these, ensuring they're consistent with any
other UI elements that want to use the same color.

In the process, I've updated the info and success accent colors to look
the same on both light and dark modes. Since these are important bits of
information, and not too bright in light mode, I think this is suitable
for now, and avoids a muddy look.

Testing Done:
Tested these in Storybook and in Review Board.

Reviewed at https://reviews.reviewboard.org/r/13728/