Ensure MenuView elements always have an ID set.
The code setting a default for MenuView's element ID didn't work
correctly. The approach failed to set an ID when the default was an
empty string, and also tried to set the ID too late (`initialize()` is
called after the element is created).

We now set the ID more explicitly during `initialize()` and during
render.

Testing Done:
Verified that a `MenuView` used in Review Board had an automatic ID
set, and that `MenuView`s with an ID didn't have it replaced.

Reviewed at https://reviews.reviewboard.org/r/13743/