Avoid inheriting some styles in ButtonGroup and Menu.
Depending on how these components are used, and what styles have been
applied elsewhere, it's possible to break the appearance of both
ButtonGroup and Menu.

ButtonGroup assumed a default margin of 0, but if anything styled
margins for a `button`-based selector, this could separate the buttons.
It was very noticeable in a MenuButton. We now hard-code an explicit
margin.

Menu items assumed a text alignment of `left`, but similarly, it would
end up inheriting an alignment. We now hard-code `left`.

Testing Done:
Verified this fixed issues with the Publish menu button in the Review
Dialog.

Reviewed at https://reviews.reviewboard.org/r/13831/