Rename design tokens for consistency.
As Ink was being built, I started honing in on ideal naming conventions.
Old names made use of "outline" for border colors, and we had a mix of
"fg" and "text color" for foreground/text elements.

Before we release Ink, this needs to change, because once we release,
it'll be more of a pain to fix the inconsistencies.

This change introduces consistent naming for the design tokens. We now
use "border-color" and "fg" naming conventions everywhere, and move
"on-dark" and "on-light" indicators to the end of the names. Going
forward, all new design tokens should follow this convention.

Testing Done:
Tested along with in-progress changes for Djblets and Review Board,
and verified that the UI still appears correct and that no CSS variables
are referencing invalid CSS variables.

Further testing will still need to be done in production.

Reviewed at https://reviews.reviewboard.org/r/13832/