Slightly lighten up the data table background.
The data table background on dark mode was hard-coded to black, which
was pretty hard on the eyes. This brings it up from #000000 to #161616,
which is easier to read and better fits the rest of the color palette in
dark mode.

Testing Done:
Tested this in the dashboard and commits list. Spent some time with it and
felt it was easier on the eyes.

Verified that the color combinations met WCAG requirements.

Reviewed at https://reviews.reviewboard.org/r/13877/