Don't open menus on MenuButtons when disabled.
`BaseMenuHandleView` now allows subclasses to specify whether a menu is
allowed to displayed, depending on the state of the component. This is
done via a `canOpenMenu()` function, which subclasses can override.

`MenuButtonView` overrides this to disable menus when the component is
disabled, ensuring it's fully non-interactive.

Testing Done:
Unit tests passed.

Tested this in Storybook.

Reviewed at https://reviews.reviewboard.org/r/13883/