Add a MenuLabel component, for label/menubar-like drop-down menu handles.
`MenuLabel` provides a simple label with an optional icon and a
drop-down handle. Hovering over it will display a drop-down menu.

This works very much like `MenuButton`, but without all the button state
options or visual appearance of a button. It's intended for simple
drop-down controls in a toolbar and for usage in menubars.

Testing Done:
Unit tests passed.

Tested in Storybook, and in usage within Review Board.

Reviewed at https://reviews.reviewboard.org/r/13884/