Add design tokens for side panels.
This adds some design tokens for standard side panels. A side panel is a
component that will be placed on the side of a container, and can be collapsed
and expanded. This is useful for things like navigation or tool menus.

In Review Board, we'll be using these definitions in the PDF navigation side
panel in the PDF review UI. The screenshots show these off, with an active
first item in the sidebar and hovered over the third item.

Testing Done:
Tested using these tokens for the PDF review UI.

Made sure that all bg colours (normal, active and hover) have WCAG AA
compliant contrast with fg colours.

Reviewed at https://reviews.reviewboard.org/r/13915/