Add Delete Item and Search icons to the icon library.
This adds icons for deleting an item (`ink-i-delete-item` -- represented
as a trash can with an "X" in it) and search (`ink-i-search` -- a
standard magnifying glass).

Testing Done:
Verified the Review Dialog was showing the correct icon.

Tested the Search icon in development code.

Reviewed at https://reviews.reviewboard.org/r/13922/