Improve readability of text on data tables on dark mode.
The data table foreground color was too bright, making it a bit hard to
read on dark mode. We now use a slightly off-white color, which improves
the readability quite a bit.

Testing Done:
Been testing locally along with a change in Review Board, and finding it
easier to read.

Verified the contrast ratio met standards and that text was easy on the
eyes.

Reviewed at https://reviews.reviewboard.org/r/13925/