Add a README and prepare package information for an Ink 0.5 release.
We now have a starting point for a README for Ink. This goes into what
Ink is, what we've built it for, what components it has, its development
status, and how to consume it. It's not comprehensive, as we still lack
documentation, but we're up-front about that, since this is a sort of
pre-release for Ink.

The package information has changed to list Beanbag as the author and to
set the version to 0.5. We also now include all of `lib` as part of the
package.

`NEWS.md` and `LICENSE` files are now included as part of the tree.

Testing Done:
Rendered the Markdown and went through it, checking for any obvious
errors. Checked spelling and links (the npm link is not live yet, though).

Built the package and inspected the contents.

Reviewed at https://reviews.reviewboard.org/r/13932/