Make -is-* classes take priority over button type.
The classes for ink buttons had combined rules for the class names and
type attributes, but this meant that it was impossible to have a
`<button type="submit" class="-is-danger">` (since the submit rule came
after the danger rule). This change splits those rules up so the classes
will take priority over the type attribute.

Testing Done:
Used this on a page that had a submit button that needed to be marked as
dangerous.

Reviewed at https://reviews.reviewboard.org/r/13974/