Fix typing on an anonymous function.
In one place inside the craft and paint implementation, we had an
anonymous function that was supposed to be typed, but the syntax was
wrong so it was defining the name of the variable but not the type.
Trivial fix.

Testing Done:
Used with other changes in ink.

Reviewed at https://reviews.reviewboard.org/r/14101/